Adds axios. Creates component folder. Adds getArticles function to the articles module.
