+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LandingPage from "./LandingPage";
+import LandingPage from "./components/LandingPage";
 
 const App = () => {
   return (
