--- src/DisplayArticles.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from "react";
-
-class DisplayArticles extends Component {
-  state = {
-    articlesList: [],
-  };
-
-  render() {
-    componentDidMount = async () => {
-      let result = await getArticles();
-      this.setState({ articlesList: result });
-    };
-
-    return <></>;
-  }
-}
-
-export default DisplayArticles;
