+++ src/modules/articles.js
@@ -1,0 +1,8 @@
+import axios from "axios";
+
+const getArticles = async () => {
+  let result = await axios.get("/articles");
+  return result.data.articles;
+};
+
+export { getArticles };
